build muondetector-gui as a debian package
With this special folder structure as in "muondetector-gui-1.0.0" it was possible to create a debian package using pbuilder. The build is "muondetector-gui-debian-package/muondetector-gui_1.0.0_amd64.deb". Try if you can install it on x64 debian (based) system. Next step: doing the same for the daemon (maybe creating a shared library) and compile them for all types of systems (32bit, arm etc.). Creating a debian package repository on our server could be interesting. Maybe if it is finished: ask a debian mentor to distribute the project worldwide?