restructuring in progress
Finally choosing a more simple, more debian(ish) way of naming the packages. Changing to the english "muon". Going to build "shared" as shared library to better suit the debian package dependency structure. There will be one package with the daemon (only available for armhf raspbian) that depends on qt-stuff and the shared library (let's call that libmuon-shared or something like this). There will be a library package and there will be the gui.