the shared library works now. Created deb packages and a windows built of the gui using the new shared library.
The creation process of symlinks in "/usr/lib/muondetector-shared" is not perfected, there will be  copies instead (this is not good). Next steps: put some graphical statistics display for the rate-counter and the satellites in the gui. Also add missing functionalities for ublox signals. Create a debian package for muondetector (also using the shared library)