fixed some bugs with file creation
Should now create data files in "/home/<user>/.muondetector-daemon/data".
If the upload works now should be checked.
There is a weird bug that files get written in the data config file twice. Maybe we do not need multiple entries in the data config file and just say all files in "data" folder are relevant. Need to think about this.