restructured filehandler, started implementing logging using lambda functions
The idea: have one generic logparameter struct and one generic signal from daemon that sends a logparameter. Then for each relevant parameter: when connecting it to the gui also connecting it to a lambda function that converts it to the generic logparameter struct and calls the logparameter signal in daemon that sends it straight to the filehandler.