Added multilateration python code
multilateration.py has a function 'tdoa_multilateration' that takes an initial guess x0 and then an array of station locations in 3D space and an array of arrival times and the speed of light. At the current state the first entry has to be the first station that gets hit by the shower. The function calculates the most probable location of the shower creation in the atmosphere.