fixed text color
When switching the theme of the OS to dark, the text often becomes white to stay visible on the dark background. For areas where the background color is fixed, this automatic change should not happen. To fix this, the text colour (css tag "color") is now manually set to black in certain places.